kashit-main/backend/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "flask-jwt-extended>=4.7.1",
    "flask-mail>=0.10.0",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "werkzeug>=3.1.3",
]
